moved date to month conversion to Utility
