+++ RUFAS/current_weather.py
@@ -1,9 +1,6 @@
 from dataclasses import dataclass
 from typing import Optional
 
-from RUFAS.time import Time
-from RUFAS.config import is_leap_year
-
 
 @dataclass
 class CurrentWeather:
@@ -80,29 +77,3 @@ class CurrentWeather:
         """
         daylength = [9, 10, 11, 13, 14, 15, 15, 15, 13, 12, 10, 9]
         return daylength[month-1]
-
-    @staticmethod
-    def day_to_month_conversion(time: Time) -> int:
-        """
-        Converts the day number into the corresponding month of the year.
-
-        Parameters
-        ----------
-        time: Time
-            Time object containing the current time of the simulation.
-
-        Returns
-        -------
-        int
-            The corresponding month of the year (1 for January, 2 for February, etc.).
-        """
-        # Define the cumulative day counts for non-leap and leap years
-        non_leap_cumulative_days_in_months = [31, 59, 90, 120, 151, 181, 212, 243, 273, 304, 334, 365]
-        leap_cumulative_days_in_months = [31, 60, 91, 121, 152, 182, 213, 244, 274, 305, 335, 366]
-
-        cumulative_days_in_months = \
-            leap_cumulative_days_in_months if is_leap_year(time.calendar_year) else non_leap_cumulative_days_in_months
-
-        for month, day_count in enumerate(cumulative_days_in_months):
-            if time.day <= day_count:
-                return month + 1
